refactor(file structure): reorganised file structure so the landing page is our worlds page
The dev folder now contains the previous landing page. I have kept it incase we want to repurpose it
into a dev portal
