+++ src/routes/+page.ts
@@ -2,7 +2,7 @@ import type { Load } from '@sveltejs/kit';
 import type { GraphData } from '$lib/types/graph';
 
 export const load: Load = async ({ fetch }) => {
-	const res = await fetch('/worlds');
+	const res = await fetch('/');
 	if (!res.ok) throw new Error('Failed to fetch graph data');
 
 	const graphData: GraphData = await res.json();
