--- src/routes/worlds/+layout.ts
@@ -1,5 +1,0 @@
-import type { LayoutLoad } from './$types';
-
-export const load: LayoutLoad = async () => {
-	return { count: 3 };
-};
